break up notifications a bit
